h8dsft_P2M2_Vincent_TD_deploy/requirements.txt
streamlit
pandas
numpy
seaborn
matplotlib
scikit-learn==1.0.2
plotly
tensorflow-cpu==2.11.0